+++ client/chat.ts
@@ -1,19 +1,25 @@
 import { h } from "snabbdom";
 
 import { _ } from './i18n';
-import RoundController from './roundCtrl';
-import AnalysisController from './analysisCtrl';
-import TournamentController from './tournament';
-import { LobbyController } from './lobby';
 import { patch } from './document';
 
-export function chatView(ctrl: RoundController | AnalysisController | TournamentController | LobbyController, chatType: string) { // TODO: instead of | better have some IChatController interface implemented by these classes
+export interface IChatController {
+    anon: boolean;
+    doSend: any;
+    spectator?: boolean;
+    gameId?: string;
+    tournamentId?: string;
+}
+
+export function chatView(ctrl: IChatController, chatType: string) { // TODO: instead of | better have some IChatController interface implemented by these classes
+    const spectator = ("spectator" in ctrl && ctrl.spectator);
+
     function onKeyPress (e: KeyboardEvent) {
         if (!(<HTMLInputElement>document.getElementById('checkbox')).checked)
             return;
         const message = (e.target as HTMLInputElement).value.trim();
         if ((e.keyCode === 13 || e.which === 13) && message.length > 0) {
-            const m: any = {type: chatType, message: message, room: ("spectator" in ctrl && ctrl.spectator) ? "spectator" : "player"};
+            const m: any = {type: chatType, message: message, room: spectator ? "spectator" : "player"};
             if ("gameId" in ctrl)
                 m["gameId"] = ctrl.gameId;
             if ("tournamentId" in ctrl)
@@ -32,7 +38,7 @@ export function chatView(ctrl: RoundController | AnalysisController | Tournament
     const anon = ctrl.anon;
     return h(`div#${chatType}.${chatType}.chat`, [
         h('div.chatroom', [
-            ((ctrl instanceof RoundController || ctrl instanceof AnalysisController) && ctrl.spectator) ? _('Spectator room') : _('Chat room'),
+            (spectator) ? _('Spectator room') : _('Chat room'),
             h('input#checkbox', { props: { title: _("Toggle the chat"), name: "checkbox", type: "checkbox", checked: "true" }, on: { click: onClick } })
         ]),
         // TODO: lock/unlock chat to spectators
