+++ client/editor.ts
@@ -3,7 +3,7 @@ import { h, VNode } from 'snabbdom';
 import { _ } from './i18n';
 import { selectVariant, VARIANTS } from './chess';
 import { EditorController } from './editorCtrl';
-import { PyChessModel } from "./main";
+import { PyChessModel } from "./types";
 
 function runEditor(vnode: VNode, model: PyChessModel) {
     const el = vnode.elm as HTMLElement;
