+++ client/gating.ts
@@ -4,8 +4,7 @@ import * as util from 'chessgroundx/util';
 import * as cg from 'chessgroundx/types';
 
 import { getPockets, lc, colorCase } from './chess';
-import RoundController from './roundCtrl';
-import AnalysisController from './analysisCtrl';
+import { GameController } from './gameCtrl';
 import { patch, bind } from './document';
 
 export interface Moves {
@@ -14,7 +13,7 @@ export interface Moves {
 }
 
 export class Gating {
-    private ctrl: RoundController | AnalysisController;
+    private ctrl: GameController;
 
     private gating : null | {
                 moves: Moves,
@@ -23,7 +22,7 @@ export class Gating {
 
     private choices: (cg.Role | "")[];
 
-    constructor(ctrl: RoundController | AnalysisController) {
+    constructor(ctrl: GameController) {
         this.ctrl = ctrl;
         this.gating = null;
         this.choices = [];
