+++ client/invite.ts
@@ -6,7 +6,7 @@ import { patch } from './document';
 import { gameType } from './profile';
 import { copyTextToClipboard } from './clipboard';
 import { timeControlStr } from './view';
-import { PyChessModel } from './main';
+import { PyChessModel } from './types';
 
 export function inviteView(model: PyChessModel): VNode[] {
     const gameId = model["gameId"];
