+++ client/roundCtrl.ts
@@ -2,9 +2,6 @@ import Sockette from 'sockette';
 
 import { h, VNode } from 'snabbdom';
 
-import * as util from 'chessgroundx/util';
-import { Chessground } from 'chessgroundx';
-import { Api } from 'chessgroundx/api';
 import * as cg from 'chessgroundx/types';
 
 import { JSONObject } from './types';
@@ -12,11 +9,9 @@ import { _, ngettext } from './i18n';
 import { patch } from './document';
 import { boardSettings } from './boardSettings';
 import { Clock } from './clock';
-import { Gating } from './gating';
-import { Promotion } from './promotion';
 import { updateMaterial } from './material';
 import { sound } from './sound';
-import { uci2LastMove, cg2uci, VARIANTS, Variant, getCounting, isHandicap, notation } from './chess';
+import { uci2LastMove, cg2uci, getCounting, isHandicap } from './chess';
 import { crosstableView } from './crosstable';
 import { chatMessage, chatView } from './chat';
 import { createMovelistButtons, updateMovelist, updateResult, selectMove } from './movelist';
@@ -24,9 +19,9 @@ import { renderRdiff } from './profile'
 import { player } from './player';
 import { updateCount, updatePoint } from './info';
 import { notify } from './notification';
-import { Clocks, MsgBoard, MsgChat, MsgFullChat, MsgGameEnd, MsgGameNotFound, MsgMove, MsgNewGame, MsgShutdown, MsgSpectators, MsgUserConnected, RDiffs, Step, CrossTable } from "./messages";
-import { PyChessModel } from "./main";
-import AnalysisController from "./analysisCtrl";
+import { Clocks, MsgBoard, MsgGameEnd, MsgMove, MsgNewGame, MsgUserConnected, RDiffs, CrossTable } from "./messages";
+import { PyChessModel } from "./types";
+import { GameController } from './gameCtrl';
 
 let rang = false;
 
@@ -43,65 +38,51 @@ interface MsgMoreTime {
 }
 
 interface MsgDrawOffer {
-	message: string;
+    message: string;
     username: string;
 }
 
 interface MsgDrawRejected {
-	message: string;
+    message: string;
 }
 
 interface MsgRematchOffer {
-	message: string;
+    message: string;
     username: string;
 }
 
 interface MsgRematchRejected {
-	message: string;
+    message: string;
 }
 
 interface MsgCount {
-	message: string;
+    message: string;
 }
 
 interface MsgSetup {
-	fen: cg.FEN;
-	color: cg.Color;
+    fen: cg.FEN;
+    color: cg.Color;
 }
 
 interface MsgGameStart {
-	gameId: string;
+    gameId: string;
 }
 
 interface MsgViewRematch {
-	gameId: string;
+    gameId: string;
 }
 
 interface MsgUpdateTV {
-	gameId: string;
+    gameId: string;
 }
 
 interface MsgBerserk {
-	color: cg.Color;
+    color: cg.Color;
 }
 
-export default class RoundController {
-    model;
-    sock;
-    chessground: Api;
-    notation: cg.Notation;
-    fullfen: string;
-    username: string;
-    anon: boolean;
-    wplayer: string;
-    bplayer: string;
-    base: number;
-    inc: number;
+export class RoundController extends GameController {
     byoyomi: boolean;
     byoyomiPeriod: number;
-    mycolor: cg.Color;
-    oppcolor: cg.Color;
-    turnColor: cg.Color;
     clocks: [Clock, Clock];
     clocktimes: Clocks;
     expirations: [VNode | HTMLElement, VNode | HTMLElement];
@@ -109,49 +90,20 @@ export default class RoundController {
     firstmovetime: number;
     tournamentGame: boolean;
     clockOn: boolean;
-    gameId: string;
-    variant: Variant;
-    chess960: boolean;
-    hasPockets: boolean;
-    vplayer0: VNode;
-    vplayer1: VNode;
-    vmaterial0: VNode;
-    vmaterial1: VNode;
+    materialDifference: boolean;
+    vmaterial0: VNode | HTMLElement;
+    vmaterial1: VNode | HTMLElement;
     vmiscInfoW: VNode;
     vmiscInfoB: VNode;
     vpng: VNode;
-    vmovelist: VNode | HTMLElement;
     vdialog: VNode;
-    gameControls: VNode;
-    moveControls: VNode;
-    ctableContainer: VNode | HTMLElement;
-    gating: Gating;
-    promotion: Promotion;
-    dests: cg.Dests; // stores all possible moves for all pieces of the player whose turn it is currently
-    promotions: string[];
-    lastmove: cg.Key[];
-    premove: {orig: cg.Key, dest: cg.Key, metadata?: cg.SetPremoveMetadata} | null;
-    predrop: {role: cg.Role, key: cg.Key} | null;
-    preaction: boolean;
-    result: string;
-    flip: boolean;
-    spectator: boolean;
     berserkable: boolean;
     settings: boolean;
     tv: boolean;
-    status: number;
-    steps: Step[];
-    pgn: string;
-    ply: number;
-    players: string[];
-    titles: string[];
-    ratings: string[];
     animation: boolean;
-    showDests: boolean; // TODO:not sure what is the point of this? doesn't chessground (especially now) have plenty of booleans like this for all kind of dests anyway?
+    showDests: boolean;
     blindfold: boolean;
     handicap: boolean;
-    autoPromote: boolean;
-    materialDifference: boolean;
     setupFen: string;
     prevPieces: cg.Pieces;
     focus: boolean;
@@ -163,6 +115,7 @@ export default class RoundController {
                           // If server received and processed it the first time, it will just ignore it
 
     constructor(el: HTMLElement, model: PyChessModel) {
+        super(el, model);
         this.focus = !document.hidden;
         document.addEventListener("visibilitychange", () => {this.focus = !document.hidden});
         window.addEventListener('blur', () => {this.focus = false});
@@ -217,104 +170,34 @@ export default class RoundController {
         const ws = (location.host.indexOf('pychess') === -1) ? 'ws://' : 'wss://';
         this.sock = new Sockette(ws + location.host + "/wsr", opts);
 
-        this.model = model;
-        this.gameId = model["gameId"] as string;
-        this.variant = VARIANTS[model["variant"]];
-        this.chess960 = model["chess960"] === 'True';
-        this.fullfen = model["fen"];
-        this.username = model["username"];
-        this.anon = model["anon"] === 'True';
-        this.wplayer = model["wplayer"];
-        this.bplayer = model["bplayer"];
-        this.base = Number(model["base"]);
-        this.inc = Number(model["inc"]);
         this.byoyomiPeriod = Number(model["byo"]);
         this.byoyomi = this.variant.timeControl === 'byoyomi';
-        this.status = Number(model["status"]);
         this.finishedGame = this.status >= 0;
         this.tv = model["tv"];
-        this.steps = [];
-        this.pgn = "";
-        this.ply = model["ply"];
 
-        this.flip = false;
         this.settings = true;
-        this.animation = localStorage.animation === undefined ? true : localStorage.animation === "true";
-        this.showDests = localStorage.showDests === undefined ? true : localStorage.showDests === "true";
         this.blindfold = localStorage.blindfold === undefined ? false : localStorage.blindfold === "true";
         this.autoPromote = localStorage.autoPromote === undefined ? false : localStorage.autoPromote === "true";
         this.materialDifference = localStorage.materialDifference === undefined ? false : localStorage.materialDifference === "true";
 
-        this.spectator = this.username !== this.wplayer && this.username !== this.bplayer;
-        this.hasPockets = this.variant.pocket;
         this.handicap = this.variant.alternateStart ? Object.keys(this.variant.alternateStart!).some(alt => isHandicap(alt) && this.variant.alternateStart![alt] === this.fullfen) : false;
 
-        this.notation = notation(this.variant);
-
-        // orientation = this.mycolor
-        if (this.spectator) {
-            this.mycolor = 'white';
-            this.oppcolor = 'black';
-        } else {
-            this.mycolor = this.username === this.wplayer ? 'white' : 'black';
-            this.oppcolor = this.username === this.wplayer ? 'black' : 'white';
-        }
-
-        // players[0] is top player, players[1] is bottom player
-        this.players = [
-            this.mycolor === "white" ? this.bplayer : this.wplayer,
-            this.mycolor === "white" ? this.wplayer : this.bplayer
-        ];
-        this.titles = [
-            this.mycolor === "white" ? this.model['btitle'] : this.model['wtitle'],
-            this.mycolor === "white" ? this.model['wtitle'] : this.model['btitle']
-        ];
-        this.ratings = [
-            this.mycolor === "white" ? this.model['brating'] : this.model['wrating'],
-            this.mycolor === "white" ? this.model['wrating'] : this.model['brating']
-        ];
-
-        this.premove = null;
-        this.predrop = null;
         this.preaction = false;
 
-        this.result = "*";
+        this.tournamentGame = this.tournamentId !== '';
         const parts = this.fullfen.split(" ");
-        this.tournamentGame = this.model["tournamentId"] !== '';
         this.clockOn = (Number(parts[parts.length - 1]) >= 2);
 
         const berserkId = (this.mycolor === "white") ? "wberserk" : "bberserk";
         // Not berserked yet, but allowed to do it
         this.berserkable = !this.spectator && this.tournamentGame && this.base > 0 && this.model[berserkId] !== 'True';
 
-        const fen_placement = parts[0];
-        this.turnColor = parts[1] === "w" ? "white" : "black";
-
-        this.steps.push({
-            'fen': this.fullfen,
-            'move': undefined,
-            'check': false,
-            'turnColor': this.turnColor,
-            });
-
-        const pocket0 = document.getElementById('pocket0') as HTMLElement;
-        const pocket1 = document.getElementById('pocket1') as HTMLElement;
-
-        this.chessground = Chessground(el, {
-            fen: fen_placement,
-            variant: this.variant.name as cg.Variant,
-            geometry: this.variant.geometry,
-            chess960: this.chess960,
-            notation: this.notation,
+        this.chessground.set({
             orientation: this.mycolor,
             turnColor: this.turnColor,
             autoCastle: this.variant.name !== 'cambodian', // TODO make more generic
             animation: { enabled: this.animation },
-
-            addDimensionsCssVars: true,
-
-            pocketRoles: this.variant.pocketRoles.bind(this.variant),
-        }, pocket0, pocket1);
+        });
 
         if (this.spectator) {
             this.chessground.set({
@@ -333,8 +216,8 @@ export default class RoundController {
                     color: this.mycolor,
                     showDests: this.showDests,
                     events: {
-                        after: this.onUserMove,
-                        afterNewPiece: this.onUserDrop,
+                        after: (orig, dest, meta) => this.onUserMove(orig, dest, meta),
+                        afterNewPiece: (role, dest, meta) => this.onUserDrop(role, dest, meta),
                     }
                 },
                 premovable: {
@@ -359,9 +242,6 @@ export default class RoundController {
             });
         }
 
-        this.gating = new Gating(this);
-        this.promotion = new Promotion(this);
-
         // initialize users
         const player0 = document.getElementById('rplayer0') as HTMLElement;
         const player1 = document.getElementById('rplayer1') as HTMLElement;
@@ -369,9 +249,9 @@ export default class RoundController {
         this.vplayer1 = patch(player1, player('player1', this.titles[1], this.players[1], this.ratings[1], model["level"]));
 
         if (this.variant.materialDifference) {
-            const material0 = document.querySelector('.material-top') as HTMLElement;
-            const material1 = document.querySelector('.material-bottom') as HTMLElement;
-            updateMaterial(this, material0, material1);
+            this.vmaterial0 = document.querySelector('.material-top') as HTMLElement;
+            this.vmaterial1 = document.querySelector('.material-bottom') as HTMLElement;
+            updateMaterial(this);
         }
 
         // initialize expirations
@@ -503,18 +383,42 @@ export default class RoundController {
 
         patch(document.getElementById('roundchat') as HTMLElement, chatView(this, "roundchat"));
 
-        boardSettings.ctrl = this;
-        const boardFamily = this.variant.board;
-        const pieceFamily = this.variant.piece;
-        boardSettings.updateBoardStyle(boardFamily);
-        boardSettings.updatePieceStyle(pieceFamily);
-        boardSettings.updateZoom(boardFamily);
-        boardSettings.updateBlindfold();
-
         this.onMsgBoard(model["board"] as MsgBoard);
     }
 
-    getGround = () => this.chessground;
+    toggleOrientation() {
+        // TODO: handle berserk
+        if (this.tournamentGame && this.ply < 2 && !this.spectator) return;
+
+        super.toggleOrientation()
+
+        boardSettings.updateDropSuggestion();
+
+        // console.log("FLIP");
+        if (this.variant.materialDifference) {
+            updateMaterial(this);
+        }
+
+        // TODO: moretime button
+        const new_running_clck = (this.clocks[0].running) ? this.clocks[1] : this.clocks[0];
+        this.clocks[0].pause(false);
+        this.clocks[1].pause(false);
+
+        const tmp_clock = this.clocks[0];
+        const tmp_clock_time = tmp_clock.duration;
+        this.clocks[0].setTime(this.clocks[1].duration);
+        this.clocks[1].setTime(tmp_clock_time);
+        if (this.status < 0) new_running_clck.start();
+
+        this.vplayer0 = patch(this.vplayer0, player('player0', this.titles[this.flip ? 1 : 0], this.players[this.flip ? 1 : 0], this.ratings[this.flip ? 1 : 0], this.model["level"]));
+        this.vplayer1 = patch(this.vplayer1, player('player1', this.titles[this.flip ? 0 : 1], this.players[this.flip ? 0 : 1], this.ratings[this.flip ? 0 : 1], this.model["level"]));
+
+        if (this.variant.counting)
+            [this.vmiscInfoW, this.vmiscInfoB] = updateCount(this.fullfen, this.vmiscInfoB, this.vmiscInfoW);
+
+        if (this.variant.materialPoint)
+            [this.vmiscInfoW, this.vmiscInfoB] = updatePoint(this.fullfen, this.vmiscInfoB, this.vmiscInfoW);
+    }
 
     private berserk = (color: cg.Color) => {
         let bclock;
@@ -582,26 +486,6 @@ export default class RoundController {
         // console.log("Resign");
         if (confirm(_('Are you sure you want to resign?'))) {
             this.doSend({ type: "resign", gameId: this.gameId });
-        }
-    }
-
-    private pass = () => {
-        let passKey: cg.Key = 'a0';
-        const pieces = this.chessground.state.pieces;
-        const dests = this.chessground.state.movable.dests!;
-        for (const [k, p] of pieces) {
-            if (p.role === 'k-piece' && p.color === this.turnColor)
-                if (dests.get(k)?.includes(k)) {
-                    passKey = k;
-                    break;
-                }
-        }
-        if (passKey !== 'a0') {
-            // prevent calling pass() again by selectSquare() -> onSelect()
-            this.chessground.state.movable.dests = undefined;
-            this.chessground.selectSquare(passKey);
-            sound.moveSound(this.variant, false);
-            this.sendMove(passKey, passKey, '');
         }
     }
 
@@ -718,11 +602,11 @@ export default class RoundController {
     }
 
     private joinTournament = () => {
-        window.location.assign(this.model["home"] + '/tournament/' + this.model["tournamentId"]);
+        window.location.assign(this.model["home"] + '/tournament/' + this.tournamentId);
     }
 
     private pauseTournament = () => {
-        window.location.assign(this.model["home"] + '/tournament/' + this.model["tournamentId"] + '/pause');
+        window.location.assign(this.model["home"] + '/tournament/' + this.tournamentId + '/pause');
     }
 
     private gameOver = (rdiffs: RDiffs) => {
@@ -1011,55 +895,26 @@ export default class RoundController {
                 }
             }
         }
-        updateMaterial(this, this.vmaterial0, this.vmaterial1);
+        updateMaterial(this);
     }
 
-    goPly = (ply: number) => {
+    goPly = (ply: number, plyVari = 0) => {
+        super.goPly(ply, plyVari);
+
         const step = this.steps[ply];
-        if (step === undefined) return;
-
-        const move = uci2LastMove(step.move);
-        let capture = false;
-        if (move.length > 0) {
-            // 960 king takes rook castling is not capture
-            // TODO Defer this logic to ffish.js
-            capture = (this.chessground.state.pieces.get(move[move.length - 1]) !== undefined && step.san?.slice(0, 2) !== 'O-') || (step.san?.slice(1, 2) === 'x');
-        }
-
         this.chessground.set({
-            fen: step.fen,
-            turnColor: step.turnColor,
             movable: {
                 free: false,
                 color: this.spectator ? undefined : step.turnColor,
                 dests: (this.turnColor === this.mycolor && this.result === "*" && ply === this.steps.length - 1) ? this.dests : undefined,
-                },
-            check: step.check,
-            lastMove: move,
+            }
         });
-        this.fullfen = step.fen;
-        updateMaterial(this, this.vmaterial0, this.vmaterial1);
 
-        if (this.variant.counting) {
-            this.updateCount(step.fen);
-        }
-
-        if (this.variant.materialPoint) {
-            this.updatePoint(step.fen);
-        }
-
-        if (ply === this.ply + 1) {
-            sound.moveSound(this.variant, capture);
-        }
-        this.ply = ply
+        updateMaterial(this);
     }
 
-    doSend = (message: JSONObject) => {
-        // console.log("---> doSend():", message);
-        this.sock.send(JSON.stringify(message));
-    }
-
-    sendMove = (orig: cg.Orig, dest: cg.Key, promo: string) => {
+    doSendMove = (orig: cg.Orig, dest: cg.Key, promo: string) => {
+        this.clearDialog();
         // pause() will add increment!
         const oppclock = !this.flip ? 0 : 1
         const myclock = 1 - oppclock;
@@ -1125,29 +980,13 @@ export default class RoundController {
         [this.vmiscInfoW, this.vmiscInfoB] = updatePoint(fen, this.vmiscInfoW, this.vmiscInfoB);
     }
 
-    private onMove = () => {
-        return (orig: cg.Key, dest: cg.Key, capturedPiece: cg.Piece) => {
-            console.log("   ground.onMove()", orig, dest, capturedPiece);
-            sound.moveSound(this.variant, !!capturedPiece);
-        }
-    }
-
-    private onDrop = () => {
-        return (piece: cg.Piece, dest: cg.Key) => {
-            // console.log("ground.onDrop()", piece, dest);
-            if (dest !== 'a0' && piece.role) {
-                sound.moveSound(this.variant, false);
-            }
-        }
-    }
-
     private setPremove = (orig: cg.Key, dest: cg.Key, metadata?: cg.SetPremoveMetadata) => {
         this.premove = { orig, dest, metadata };
         // console.log("setPremove() to:", orig, dest, meta);
     }
 
     private unsetPremove = () => {
-        this.premove = null;
+        this.premove = undefined;
         this.preaction = false;
     }
 
@@ -1157,7 +996,7 @@ export default class RoundController {
     }
 
     private unsetPredrop = () => {
-        this.predrop = null;
+        this.predrop = undefined;
         this.preaction = false;
     }
 
@@ -1166,64 +1005,12 @@ export default class RoundController {
         // TODO: promotion?
         // console.log("performPremove()", orig, dest, meta);
         this.chessground.playPremove();
-        this.premove = null;
     }
 
     private performPredrop = () => {
         // const { role, key } = this.predrop;
         // console.log("performPredrop()", role, key);
         this.chessground.playPredrop();
-        this.predrop = null;
-    }
-
-    private onUserMove = (orig: cg.Key, dest: cg.Key, meta: cg.MoveMetadata) => {
-        onUserMove(this, orig, dest, meta);
-        this.clearDialog();
-    }
-
-    private onUserDrop = (role: cg.Role, dest: cg.Key, meta: cg.MoveMetadata) => {
-        onUserDrop(this, role, dest, meta);
-    }
-
-    private onSelect = () => {
-        let lastTime = performance.now();
-        let lastKey: cg.Key = 'a0';
-        return (key: cg.Key) => {
-            if (this.chessground.state.movable.dests === undefined) return;
-
-            const curTime = performance.now();
-
-            // Save state.pieces to help recognise 960 castling (king takes rook) moves
-            // Shouldn't this be implemented in chessground instead?
-            if (this.chess960 && this.variant.gate) {
-                this.prevPieces = new Map(this.chessground.state.pieces);
-            }
-
-            // Sittuyin in place promotion on double click
-            if (lastKey === key && curTime - lastTime < 500) {
-                if (this.chessground.state.movable.dests.get(key)?.includes(key)) {
-                    const piece = this.chessground.state.pieces.get(key)!;
-                    if (this.variant.name === 'sittuyin') { // TODO make this more generic
-                        // console.log("Ctrl in place promotion", key);
-                        const pieces: cg.Pieces = new Map();
-                        pieces.set(key, {
-                            color: piece.color,
-                            role: 'f-piece',
-                            promoted: true
-                        });
-                        this.chessground.setPieces(pieces);
-                        this.chessground.state.movable.dests = undefined;
-                        this.chessground.selectSquare(key);
-                        sound.moveSound(this.variant, false);
-                        this.sendMove(key, key, 'f');
-                    }
-                }
-                lastKey = 'a0';
-            } else {
-                lastKey = key;
-                lastTime = curTime;
-            }
-        }
     }
 
     private renderExpiration = () => {
@@ -1282,11 +1069,6 @@ export default class RoundController {
         }
     }
 
-    private onMsgSpectators = (msg: MsgSpectators) => {
-        const container = document.getElementById('spectators') as HTMLElement;
-        patch(container, h('under-left#spectators', _('Spectators: ') + msg.spectators));
-    }
-
     private onMsgUserPresent = (msg: MsgUserPresent) => {
         // console.log(msg);
         if (msg.username === this.players[0]) {
@@ -1307,24 +1089,6 @@ export default class RoundController {
             const container = document.getElementById('player1') as HTMLElement;
             patch(container, h('i-side.online#player1', {class: {"icon": true, "icon-online": false, "icon-offline": true}}));
         }
-    }
-
-    private onMsgChat = (msg: MsgChat) => {
-        if ((this.spectator && msg.room === 'spectator') || (!this.spectator && msg.room !== 'spectator') || msg.user.length === 0) {
-            chatMessage(msg.user, msg.message, "roundchat", msg.time);
-        }
-    }
-
-    private onMsgFullChat = (msg: MsgFullChat) => {
-        // To prevent multiplication of messages we have to remove old messages div first
-        patch(document.getElementById('messages') as HTMLElement, h('div#messages-clear'));
-        // then create a new one
-        patch(document.getElementById('messages-clear') as HTMLElement, h('div#messages'));
-        msg.lines.forEach((line) => {
-            if ((this.spectator && line.room === 'spectator') || (!this.spectator && line.room !== 'spectator') || line.user.length === 0) {
-                chatMessage(line.user, line.message, "roundchat", line.time);
-            }
-        });
     }
 
     private onMsgMoreTime = (msg: MsgMoreTime) => {
@@ -1360,15 +1124,6 @@ export default class RoundController {
         this.clearDialog();
     }
 
-    private onMsgGameNotFound = (msg: MsgGameNotFound) => {
-        alert(_("Requested game %1 not found!", msg['gameId']));
-        window.location.assign(this.model["home"]);
-    }
-
-    private onMsgShutdown = (msg: MsgShutdown) => {
-        alert(msg.message);
-    }
-
     private onMsgCount = (msg: MsgCount) => {
         chatMessage("", msg.message, "roundchat");
         if (msg.message.endsWith("started")) {
@@ -1385,7 +1140,7 @@ export default class RoundController {
         }
     }
 
-    private onMessage = (evt: MessageEvent) => {
+    protected onMessage(evt: MessageEvent) {
         // console.log("<+++ onMessage():", evt.data);
         const msg = JSON.parse(evt.data);
         switch (msg.type) {
@@ -1404,17 +1159,8 @@ export default class RoundController {
             case "user_present":
                 this.onMsgUserPresent(msg);
                 break;
-            case "spectators":
-                this.onMsgSpectators(msg);
-                break
             case "user_disconnected":
                 this.onMsgUserDisconnected(msg);
-                break;
-            case "roundchat":
-                this.onMsgChat(msg);
-                break;
-            case "fullchat":
-                this.onMsgFullChat(msg);
                 break;
             case "new_game":
                 this.onMsgNewGame(msg);
@@ -1440,15 +1186,6 @@ export default class RoundController {
             case "updateTV":
                 this.onMsgUpdateTV(msg);
                 break
-            case "game_not_found":
-                this.onMsgGameNotFound(msg);
-                break
-            case "shutdown":
-                this.onMsgShutdown(msg);
-                break;
-            case "logout":
-                this.doSend({type: "logout"});
-                break;
             case "setup":
                 this.onMsgSetup(msg);
                 break;
@@ -1460,70 +1197,4 @@ export default class RoundController {
                 break;
         }
     }
-}
-
-/**
- * Custom variant-specific logic to be triggered on move and alter state of board/pocket depending on variant rules.
- * TODO: contains also some ui logic - maybe good to split pure chess rules (which maybe can go to chess.ts?)
- *       from rendering dialogs and
- * TODO: Unify this with analysisCtrl
- * */
-export function onUserMove(ctrl: RoundController | AnalysisController, orig: cg.Key, dest: cg.Key, meta: cg.MoveMetadata) {
-        ctrl.preaction = meta.premove;
-        // chessground doesn't knows about ep, so we have to remove ep captured pawn
-        const pieces = ctrl.chessground.state.pieces;
-        // console.log("ground.onUserMove()", orig, dest, meta);
-        let moved = pieces.get(dest);
-        // Fix king to rook 960 castling case
-        if (moved === undefined) moved = {role: 'k-piece', color: ctrl.mycolor} as cg.Piece;
-        if (meta.captured === undefined && moved !== undefined && moved.role === "p-piece" && orig[0] !== dest[0] && ctrl.variant.enPassant) {
-            const pos = util.key2pos(dest),
-            pawnPos: cg.Pos = [pos[0], pos[1] + (ctrl.mycolor === 'white' ? -1 : 1)];
-            const diff: cg.PiecesDiff = new Map();
-            diff.set(util.pos2key(pawnPos), undefined);
-            ctrl.chessground.setPieces(diff);
-            meta.captured = {role: "p-piece", color: moved.color === "white"? "black": "white"/*or could get it from pieces[pawnPos] probably*/};
-        }
-        // increase pocket count
-        // important only during gap before we receive board message from server and reset whole FEN (see also onUserDrop)
-        if (ctrl.variant.drop && meta.captured) {
-            let role = meta.captured.role;
-            if (meta.captured.promoted)
-                role = (ctrl.variant.promotion === 'shogi' || ctrl.variant.promotion === 'kyoto') ? meta.captured.role.slice(1) as cg.Role : "p-piece";
-
-            const pocket = ctrl.chessground.state.pockets ? ctrl.chessground.state.pockets[util.opposite(meta.captured.color)] : undefined;
-            if (pocket && role && role in pocket) {
-                pocket[role]!++;
-                ctrl.chessground.state.dom.redraw(); // TODO: see todo comment also at same line in onUserDrop.
-            }
-        }
-
-        //  gating elephant/hawk
-        if (ctrl.variant.gate) {
-            if (!ctrl.promotion.start(moved.role, orig, dest, meta.ctrlKey) && !ctrl.gating.start(ctrl.fullfen, orig, dest)) ctrl.sendMove(orig, dest, '');
-        } else {
-            if (!ctrl.promotion.start(moved.role, orig, dest, meta.ctrlKey)) ctrl.sendMove(orig, dest, '');
-            ctrl.preaction = false;
-        }
-}
-
-/**
- * Variant specific logic for when dropping a piece from pocket is performed
- * todo: decreasing of pocket happens here as well even though virtually no variant ever has a drop rule that doesn't decrease pocket.
- *       Only reason currently this is not in chessground is editor where we have a second "pocket" that serves as a palette
- *       Also maybe nice ot think if ui+communication logic can be split out of here (same for onUserMove) so only chess rules remain?
- * */
-export function onUserDrop(ctrl: RoundController | AnalysisController, role: cg.Role, dest: cg.Key, meta: cg.MoveMetadata) {
-    ctrl.preaction = meta.predrop === true;
-    // decrease pocket count - todo: covers the gap before we receive board message confirming the move - then FEN is set
-    //                               and overwrites whole board+pocket and refreshes.
-    //                               Maybe consider decrease count on start of drag (like in editor mode)?
-    ctrl.chessground.state.pockets![ctrl.chessground.state.turnColor]![role]! --;
-    ctrl.chessground.state.dom.redraw();
-    if (ctrl.variant.promotion === 'kyoto') {
-        if (!ctrl.promotion.start(role, 'a0', dest)) ctrl.sendMove(util.dropOrigOf(role), dest, '');
-    } else {
-        ctrl.sendMove(util.dropOrigOf(role), dest, '')
-    }
-    ctrl.preaction = false;
 }
